First draft of AutoAPI functionality
This commit adds functionality that, given a project root and a set of patterns
(strings or globs) to ignore, identifies all Python files in the root that don't
match any of the excluded globs and generates Markdown files that `mkdocstrings`
can use to produce HTML documentation for them.

There is surely room for improvement, but as a first draft this works nicely.