Apply updates to setup.py
This commit updates the version of the package (had a couple errors during
testing so it's higher than 0.1.1) and a couple of links (accidentally forgot to
edit the links copy-pasted from `django-postgres-loader`). In addition, it adds
an entry point which is required for `MkDocs` to recognize the package as a
plugin.