Add documentation for navigation configuration to README (#2)
I received some feedback that it was not obvious how to include auto-generated
API documentation in a site's navigation; this commit fixes that by adding a
section explaining how to include the API docs via the `nav` setting in
`mkdocs.yml`. I also added a section showing a simple full `mkdocs.yml` for a
project.