Only generate local Markdown on mkdocs build (#11)
This commit adds a startup event hook to the plugin which adds the command used
(`build`, `gh-deploy`, or `serve`) to the configuration object, and updates the
API documentation file generation function so that it only builds local copies
of the files if the command is `build`. I've confirmed that this fixes the bug
described in #11.