Add support for VBA handler (#53)
This turned out to be rather simple; just had to get the `mkdocstrings` default
handler configuration and create the module identifier for each file based on
its value (if `python` then `path.to.module`, if `vba` then
`path/to/module.bas`). I also added validation of the default handler to make
sure it is supported.