Address issue with navigation when using mkdocs theme (#7)
I've observed that, when using the mkdocs theme, the "literate nav" capability
(where clicking the name of a directory in the navigation links to documentation
for its `__init__.py` file) does not work as expected. To get around this, I
added logic that creates an "Index" page within the directory if the theme is
mkdocs; the literate nav capability still works as expected for the readthedocs
and material themes.