Fix bug causing crash when `nav` undefined but `autoapi_add_nav_entry` not (#70)
Had a user report a crash when attempting to run `mkdocs build` on their
project. Turns out that attempting to add a new nav entry via
`autoapi_add_nav_entry` when there is no `nav` config in `mkdocs.yml` is the
cause. This commit addresses the bug by setting the nav to an empty list before
adding the new entry if the `nav` config is `None`. I also have added a
changelog entry and incremented the project version to 0.4.1 to reflect the
bugfix.